+++ changelog.md
@@ -1,2 +1,2 @@
 ### Rechiseled: Create 1.0.0
-- Initial release of Rechiseled: Create
+- Initial release of Rechiseled: Create for Fabric on 1.20